+++ apps/demos/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fluid-ds/demos
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @fluid-ds/components@0.1.4
+
 ## 0.0.4
 
 ### Patch Changes